replace ddd endpoint with ddd2
